Engine.cpp as separate files
